// src/models/IMovieDetails.ts
export interface IMovieDetails {
	Title: string;
	Year: string;
	Rated: string;
	Released: string;
	Runtime: string;
	Genre: string;
	Director: string;
	Writer: string;
	Actors: string;
	Plot: string;
	Language: string;
	Country: string;
	Production: string;
	Awards: string;
	Poster: string;
	imdbID: string;
	Response: string;
	DVD: string;
}